Petitions now need an image to be submitted, and the saved image is saved to the petition
to add a petition, an file must be submitted with it (no error shown for this yet though). Then, it first creates the petition, and then saves  the image to that petition, before redirecting to its page.
ISSUE: if a file is uploaded thats not an image, the petition is saved but the image is not