Updated petitions list dialogs to not apply unless directly confirmed.
This fixes a previous bug where entering a query into a dialog, closing it, then confirming a different query, would still apply the first query, etc.
This was fixed by making the dialogs only update temporary state variables, until submit is confirmed, upon which, it will be applied to the actual state variables.