Pagination implemented, but still need to allow changing page size
pagination implemented, but currently only works with the default value of 10 petitions per page. However, the code is very unoptimized, im very sleepy. Namely, having both getPetitions and fetchPetitionsWithPagination, as well as the different useEffect added. as well as that I would like to update the page switching as well.